+++ R/node_types.R
@@ -370,14 +370,14 @@ variable_node <- R6Class(
 
       # sum across all dimensions of jacobian
       already_summed <-
-        identical(dim(ljd), list(NULL)) | identical(dim(ljd), list())
+        identical(dim(ljd), NA_integer_) | identical(dim(ljd), integer(0))
 
       if (!already_summed) {
         ljd <- tf_sum(ljd, drop = TRUE)
       }
 
       # make sure there's something in the batch dimension
-      if (identical(dim(ljd), list())) {
+      if (identical(dim(ljd), integer(0))) {
         ljd <- tf$expand_dims(ljd, 0L)
       }