+++ R/test_if_forked_cluster.R
@@ -1,19 +1,12 @@
 test_if_forked_cluster <- function(){
-  dummy <- parallelly::isForkedChild()
-
-  f <- future({
-    if (parallelly::isForkedChild()) {
-      msg <- cli::format_error(
-        "parallel mcmc samplers cannot be run with a fork cluster"
-      )
-      stop(
-        msg,
-        call. = FALSE
-      )
-    }
-
-    42
-  })
-
-  value(f)
+  is_forked <- value(future(parallelly::isForkedChild()))
+  if (is_forked) {
+    msg <- cli::format_error(
+      "parallel mcmc samplers cannot be run with a fork cluster"
+    )
+    stop(
+      msg,
+      call. = FALSE
+    )
+  }
 }